test: split diplomatic order-suggestion test under 400-line domain gate (#3758)
Move minor-nation/overture suggestDiplomaticOrders cases into a sibling
order_suggestion_api_impl_diplomatic_minor_test.dart so each file stays at
or below the repo.domain_package_test_file_size 400-line cap. No behavior
change; all 10 cases preserved.

Refs #3758
